requirements.txt
# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Google Gemini AI
google-genai>=1.0.0

# GUI
PyQt6>=6.5.0

# Utilities
Pillow>=10.0.0